Fix error in file repository when resetting

+++ lib/src/pond/repository/file/with_file_entity_repository.dart
@@ -59,7 +59,9 @@ mixin WithFileEntityRepository on EntityRepository implements WithCacheEntityRep
   Future<void> onReset(AppContext context) async {
     await super.onReset(context);
 
-    await _baseDirectory.delete(recursive: true);
+    if (await _baseDirectory.exists()) {
+      await _baseDirectory.delete(recursive: true);
+    }
   }
 
   File _getFile(String id) {
